feat: integrate react-router-dom for navigation and add Admin Dashboard
