+++ src/constants.ts
@@ -2,5 +2,5 @@
  * Base URL for the API.
  * Update this value to switch between environments (e.g., development, production).
  */
-// export const API_BASE_URL = 'http://localhost:8080'; // Local development URL
-export const API_BASE_URL = 'https://th-backend-kko4.onrender.com' // Production URL
+export const API_BASE_URL = 'http://localhost:8080'; // Local development URL
+// export const API_BASE_URL = 'https://th-backend-kko4.onrender.com' // Production URL
